api banner

+++ app/Http/Controllers/FileController.php
@@ -2,7 +2,8 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\ImageModel;
+use App\Models\BannerImageModel;
+use App\Models\ProductImageModel;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 // use Illuminate\Support\Facades\File; 
@@ -10,6 +11,16 @@ use Illuminate\Support\Facades\Storage;
 
 class FileController extends Controller
 {
+    public function getAllDataProductImg()
+    {
+        return response()->json(ProductImageModel::get(), 200);
+    }
+
+    public function getAllDataBannerImg()
+    {
+        return response()->json(BannerImageModel::orderByRaw('priority ASC')->get(), 200);
+    }
+
     public function getImgByName($fileName)
     {
         return response()->download(public_path('img/'.$fileName));
@@ -18,16 +29,23 @@ class FileController extends Controller
     public function getImgByProductDetailId($product_detail_id)
     {
         //can get only 1 image with 1 product_detail_id
-        $list_image = ImageModel::where('product_detail_id', $product_detail_id)->get();
+        $list_image = ProductImageModel::where('product_detail_id', $product_detail_id)->get();
         return response()->download(public_path('img/'.$list_image->first()->name));
     }
 
-    public function uploadImg(Request $request)
+    public function getImgByPriority($priority)
+    {
+        //can get only 1 image with 1 product_detail_id
+        $image = ProductImageModel::where('priority', $priority)->get();
+        return response()->download(public_path('img/'.$image->first()->name));
+    }
+
+    public function uploadProductImg(Request $request)
     {
         $fileName = Str::random(10).'.png';
 
         //1. save link image in database
-        $image = new ImageModel();
+        $image = new ProductImageModel();
         $image->name = $fileName;
         $image->origin_name = $request->file('photo')->getClientOriginalName();
         $image->product_detail_id = $request->product_detail_id;
@@ -40,9 +58,27 @@ class FileController extends Controller
         return response()->json(['url: ', $photoURL], 200);
     }
 
-    public function deleteImg($id)
+    public function uploadBannerImg(Request $request)
     {
-        $image = ImageModel::find($id);
+        $fileName = Str::random(10).'.png';
+
+        //1. save link image in database
+        $image = new BannerImageModel();
+        $image->name = $fileName;
+        $image->origin_name = $request->file('photo')->getClientOriginalName();
+        $image->priority = $request->priority;
+        $image->save();
+
+        //2. storage image in file system public/img/..
+        $path = $request->file('photo')->move(public_path('/img'), $fileName);
+        $photoURL = url('/'.$fileName);
+
+        return response()->json(['url: ', $photoURL], 200);
+    }
+
+    public function deleteProductImgById($id)
+    {
+        $image = ProductImageModel::find($id);
         
         if (is_null($image))
         {
@@ -60,9 +96,9 @@ class FileController extends Controller
         return response()->json(['Deleted'], 200);
     }
 
-    public function deleteImgByProductDetail($product_detail_id)
+    public function deleteProductImgByProductDetail($product_detail_id)
     {
-        $image = ImageModel::where('product_detail_id', $product_detail_id)->get();
+        $image = ProductImageModel::where('product_detail_id', $product_detail_id)->get();
         
         //check image in database
         if (empty(json_decode($image, true)))
@@ -75,6 +111,26 @@ class FileController extends Controller
             Storage::disk('local_public')->delete($image->name);
         }
         else return response()->json(["message" => "Image Not Found"], 404);
+
+        //delete fil name in database
+        $image->delete();
+
+        return response()->json(['Deleted'], 200);
+    }
+
+    public function deleteBannerImgById($id)
+    {
+        $image = BannerImageModel::find($id);
+        
+        if (is_null($image))
+        {
+            return response()->json(["message" => "ID Not Found"], 404);
+        }
+
+        //delete file in folder
+        if(Storage::disk('local_public')->exists($image->name)){
+            Storage::disk('local_public')->delete($image->name);
+        }
 
         //delete fil name in database
         $image->delete();
